requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
choreographer==1.0.9
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
durationpy==0.10
fastapi==0.116.1
filelock==3.18.0
flasgger==0.9.7.1
Flask==3.1.1
flask-sock==0.7.0
flatbuffers==25.2.10
fsspec==2025.7.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kaleido==1.0.0
kubernetes==33.1.0
logistro==1.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistune==3.1.3
mmh3==5.1.0
mpmath==1.3.0
narwhals==1.47.0
numpy==2.3.1
oauthlib==3.3.1
ollama==0.5.1
onnxruntime==1.22.1
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-instrumentation-asgi==0.56b0
opentelemetry-instrumentation-fastapi==0.56b0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opentelemetry-util-http==0.56b0
orjson==3.11.0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
plotly==6.2.0
posthog==6.1.0
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
shellingham==1.5.4
simple-websocket==1.1.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tokenizers==0.21.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vanna==0.7.9
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
zipp==3.23.0
